refactor: add smi_to_srt_converter
